web0003_陈新尧

+++ Models/Room.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int DId { get; set; }
         public int Num { get; set; }
-        public int Margin { get; set; }
+        public byte Margin { get; set; }
         public int Price { get; set; }
     }
 }
